chore: rename lib/base/log_outputters => lib/base/LogOutputters
