+++ src/lib/base/Log.cpp
@@ -7,7 +7,7 @@
 
 #include "base/Log.h"
 #include "arch/Arch.h"
-#include "base/log_outputters.h"
+#include "base/LogOutputters.h"
 #include "common/constants.h"
 
 #include <cstdarg>
