+++ src/lib/platform/MSWindowsWatchdog.cpp
@@ -10,8 +10,8 @@
 #include "arch/win32/XArchWindows.h"
 #include "base/ELevel.h"
 #include "base/Log.h"
+#include "base/LogOutputters.h"
 #include "base/TMethodJob.h"
-#include "base/log_outputters.h"
 #include "deskflow/App.h"
 #include "mt/Thread.h"
 #include "platform/MSWindowsHandle.h"
